// smartcoin-frontend/src/app/api/model/accountListResponse.ts
/**
 * Smartcoin API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CurrencySubtotal } from './currencySubtotal';
import { AccountResponse } from './accountResponse';


/**
 * Las cuentas del usuario con su saldo actual y un subtotal por moneda.
 */
export interface AccountListResponse { 
    /**
     * Ordenadas por moneda y nombre.
     */
    accounts: Array<AccountResponse>;
    /**
     * Un subtotal por cada moneda que tiene cuentas, en el orden ARS, USD. Nunca hay un total que mezcle monedas (RN-04).
     */
    subtotals: Array<CurrencySubtotal>;
}